refactor(kimi-k2): split runner bringup from schedulers
