+++ pegainfer-kimi-k2/src/runner.rs
@@ -1,10 +1,10 @@
 mod affinity;
+mod bringup;
 mod config;
-mod engine;
 mod executor;
 mod load_balancer;
 mod moe_pplx;
 mod scheduler;
 mod worker;
 
-pub(crate) use scheduler::start_engine;
+pub(crate) use bringup::start_engine;
